Projeto/requirements.txt
gradio>=4.44.1
PyPDF2==3.0.1
transformers>=4.38.0
torch>=2.1.0
accelerate>=0.27.2
sentencepiece
faiss-cpu
numpy
sentence-transformers
